Move QgsLabelPosition out to own file
